Add files via upload
So I did a bunch more code, the GUI can make accompanists, soloists with an array of compositions/date last played, and conductors that have an array of concerts.  I'll make a new branch but you should commit it to the master branch if we want to use this as the master.  I'll try and integrate some of Claires classes tomorrow.